Locking down comments and posts.

+++ src/resolvers/Mutation.js
@@ -21,18 +21,20 @@ const Mutation = {
             token: jwt.sign({userId: user.id}, 'thisisasecret')
         }
     },
-    updateUser(parent, args, { prisma }, info){
+    updateUser(parent, args, { prisma, request }, info){
+        const userId = getUserId(request);
         return prisma.mutation.updateUser({
             where: {
-                id: args.id
+                id: userId
             },
             data: args.data
         })
     },
-    deleteUser(parent, args, { prisma }, info) {
+    deleteUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request);
         return prisma.mutation.deleteUser({
             where: {
-                id: args.id
+                id: userId
             }
         }, info)
     },
@@ -52,7 +54,19 @@ const Mutation = {
             }
         }, info)
     },
-    updatePost(parent, args, { prisma }, info){
+    async updatePost(parent, args, { prisma, request }, info){
+        const userId = getUserId(request);
+
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!postExists) {
+            throw new Error('Unable to update post')
+        }
         return prisma.mutation.updatePost({
             data: args.data,
             where: {
@@ -60,7 +74,20 @@ const Mutation = {
             }
         }, info)
     },
-    deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request);
+
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!postExists) {
+            throw new Error('Unable to delete post')
+        }
+
         return prisma.mutation.deletePost({
             where: {
                 id: args.id
@@ -68,13 +95,15 @@ const Mutation = {
         }, info)
     },
 
-    createComment(parent, args, { prisma }, info){
+    createComment(parent, args, { prisma, request }, info){
+        const userId = getUserId(request);
+
         return prisma.mutation.createComment({
             data: {
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -85,7 +114,20 @@ const Mutation = {
             }
         }, info)
     },
-    updateComment(parent, args, { prisma }, info) {
+    async updateComment(parent, args, { prisma, request}, info) {
+        const userId = getUserId(request);
+
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!commentExists) {
+            throw new Error ('Unable to update comment.')
+        }
+
         return prisma.mutation.updateComment ({
             where: {
                 id: args.id
@@ -93,7 +135,20 @@ const Mutation = {
             data: args.data
         }, info)
     },
-    deleteComment(parent, args, { prisma }, info) {
+    async deleteComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request);
+
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!commentExists) {
+            throw new Error ('Unable to delete comment.')
+        }
+
         return prisma.mutation.deleteComment({
             where: {
                 id: args.id
